64ビットLinux環境でのワーニングを抑止

+++ graphic/scenes/table/showdice.cpp
@@ -10,12 +10,12 @@ using utils::playerRelative;
 /* サイコロを表示する */
 void GameTableScreen::DiceReconst::ShowDice(const GameTable* gameStat) {
 	RECT rect1 = {
-		static_cast<long>((DiceWidth + DicePadding) * (gameStat->Dice[0].Number - 1)), static_cast<long>((DiceHeight + DicePadding) * (gameStat->Dice[0].Direction    )),
-		static_cast<long>((DiceWidth + DicePadding) * (gameStat->Dice[0].Number    )), static_cast<long>((DiceHeight + DicePadding) * (gameStat->Dice[0].Direction + 1)),
+		static_cast<int32_t>((DiceWidth + DicePadding) * (gameStat->Dice[0].Number - 1)), static_cast<int32_t>((DiceHeight + DicePadding) * (gameStat->Dice[0].Direction    )),
+		static_cast<int32_t>((DiceWidth + DicePadding) * (gameStat->Dice[0].Number    )), static_cast<int32_t>((DiceHeight + DicePadding) * (gameStat->Dice[0].Direction + 1)),
 	};
 	RECT rect2 = {
-		static_cast<long>((DiceWidth + DicePadding) * (gameStat->Dice[1].Number - 1)), static_cast<long>((DiceHeight + DicePadding) * (gameStat->Dice[1].Direction    )),
-		static_cast<long>((DiceWidth + DicePadding) * (gameStat->Dice[1].Number    )), static_cast<long>((DiceHeight + DicePadding) * (gameStat->Dice[1].Direction + 1)),
+		static_cast<int32_t>((DiceWidth + DicePadding) * (gameStat->Dice[1].Number - 1)), static_cast<int32_t>((DiceHeight + DicePadding) * (gameStat->Dice[1].Direction    )),
+		static_cast<int32_t>((DiceWidth + DicePadding) * (gameStat->Dice[1].Number    )), static_cast<int32_t>((DiceHeight + DicePadding) * (gameStat->Dice[1].Direction + 1)),
 	};
 	switch (playerRelative(gameStat->GameRound % Players, gameStat->PlayerID)) {
 	case sSelf:
